Feature settings integration

+++ server/mentionscrawler/blueprints/company.py
@@ -1,7 +1,7 @@
-from flask import Blueprint, request
+from flask import Blueprint, request, jsonify
 from ..models.company import Company
-from ..db import insert_row
-from ..responses import bad_request_response, ok_response
+from ..db import insert_rows
+from ..responses import ok_response, bad_request_response
 from ..authentication.authenticate import authenticate, enforce_json
 
 company_bp = Blueprint("companies", __name__, url_prefix="/")
@@ -11,12 +11,14 @@ company_bp = Blueprint("companies", __name__, url_prefix="/")
 @company_bp.route("/companies", methods=["PUT"])
 @enforce_json()
 @authenticate()
-def companies(user):
-    body = request.get_json()
-    if body.get("names"):
-        names = body.get("names")
-        Company.query.delete()
-        for name in names:
-            insert_row(Company(user.get("user_id"), name))
-        return ok_response("Company names updated!")
-    return bad_request_response("Invalid request! Missing fields!")
+def update_companies(user):
+    names = request.get_json()
+    if len(names) == 0:
+        return bad_request_response("Must have at least one company name")
+    Company.query.filter_by(mention_user_id = user.get("user_id")).delete()
+    companies = []
+    for name in names:
+        companies.append(Company(user.get("user_id"), name))
+    insert_rows(companies)
+    return ok_response("Company names updated!")
+
